requirements.txt
tensorflow-gpu
tensorflow
numpy
scipy
pillow
imageio
pretty_midi
pypianoroll
pyyaml
